# docs/openapi-mvp-skeleton-draft.yaml
openapi: 3.1.0
info:
  title: Order System v2 API
  version: 0.1.0
paths:
...
  /api/v1/allocations/{allocation_id}/override:
    patch:
      tags:
      - allocations
      summary: Override Allocation
      operationId: override_allocation_api_v1_allocations__allocation_id__override_patch
      security:
      - HTTPBearer: []
      parameters:
      - name: allocation_id
        in: path
        required: true
        schema:
          type: integer
          title: Allocation Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AllocationOverrideRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AllocationResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '409':
          description: Conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
  /api/v1/allocations/{allocation_id}/split-line:
    post:
      tags:
      - allocations
      summary: Split Allocation
      operationId: split_allocation_api_v1_allocations__allocation_id__split_line_post
      security:
      - HTTPBearer: []
      parameters:
      - name: allocation_id
        in: path
        required: true
        schema:
          type: integer
          title: Allocation Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AllocationSplitRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AllocationResponse'
                title: Response Split Allocation Api V1 Allocations  Allocation Id  Split
                  Line Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '409':
          description: Conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
...
components:
  schemas:
    AllocationOverrideRequest:
      properties:
        final_supplier_id:
          type: integer
          exclusiveMinimum: 0.0
          title: Final Supplier Id
        final_qty:
          type: number
          minimum: 0.0
          title: Final Qty
        final_uom:
          type: string
          maxLength: 32
          minLength: 1
          title: Final Uom
        override_reason_code:
          type: string
          maxLength: 64
          minLength: 1
          title: Override Reason Code
        override_note:
          anyOf:
          - type: string
            maxLength: 1000
          - type: 'null'
          title: Override Note
      type: object
      required:
      - final_supplier_id
      - final_qty
      - final_uom
      - override_reason_code
      title: AllocationOverrideRequest
    AllocationResponse:
      properties:
        id:
          type: integer
          title: Id
        order_item_id:
          type: string
          title: Order Item Id
        suggested_supplier_id:
          anyOf:
          - type: integer
          - type: 'null'
          title: Suggested Supplier Id
        suggested_qty:
          anyOf:
          - type: number
          - type: 'null'
          title: Suggested Qty
        final_supplier_id:
          anyOf:
          - type: integer
          - type: 'null'
          title: Final Supplier Id
        final_qty:
          anyOf:
          - type: number
          - type: 'null'
          title: Final Qty
        final_uom:
          anyOf:
          - type: string
          - type: 'null'
          title: Final Uom
        is_manual_override:
          type: boolean
          title: Is Manual Override
        override_reason_code:
          anyOf:
          - type: string
          - type: 'null'
          title: Override Reason Code
        target_price:
          anyOf:
          - type: number
          - type: 'null'
          title: Target Price
        split_group_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Split Group Id
        parent_allocation_id:
          anyOf:
          - type: integer
          - type: 'null'
          title: Parent Allocation Id
        is_split_child:
          type: boolean
          title: Is Split Child
        created_at:
          type: string
          format: date-time
          title: Created At
        updated_at:
          type: string
          format: date-time
          title: Updated At
      type: object
      required:
      - id
      - order_item_id
      - suggested_supplier_id
      - suggested_qty
      - final_supplier_id
      - final_qty
      - final_uom
      - is_manual_override
      - override_reason_code
      - target_price
      - split_group_id
      - parent_allocation_id
      - is_split_child
      - created_at
      - updated_at
      title: AllocationResponse
...
    AllocationSplitRequest:
      properties:
        parts:
          items:
            $ref: '#/components/schemas/SplitPart'
          type: array
          minItems: 2
          title: Parts
        override_reason_code:
          type: string
          maxLength: 64
          minLength: 1
          title: Override Reason Code
        override_note:
          anyOf:
          - type: string
            maxLength: 1000
          - type: 'null'
          title: Override Note
      type: object
      required:
      - parts
      - override_reason_code
      title: AllocationSplitRequest
...
    ApiErrorDetail:
      properties:
        code:
          type: string
          title: Code
        message:
          type: string
          title: Message
        details:
          anyOf:
          - items:
              type: object
            type: array
          - type: 'null'
          title: Details
      type: object
      required:
      - code
      - message
      title: ApiErrorDetail
    ApiErrorResponse:
      properties:
        detail:
          $ref: '#/components/schemas/ApiErrorDetail'
      type: object
      required:
      - detail
      title: ApiErrorResponse
...
    SplitPart:
      properties:
        final_supplier_id:
          type: integer
          exclusiveMinimum: 0.0
          title: Final Supplier Id
        final_qty:
          type: number
          minimum: 0.0
          title: Final Qty
        final_uom:
          type: string
          maxLength: 32
          minLength: 1
          title: Final Uom
      type: object
      required:
      - final_supplier_id
      - final_qty
      - final_uom
      title: SplitPart
...
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer